Check for ssh auth sockets in xdg dir.
Also now blindly setting those as our sock instead of lsof'ing.

The reason for the blind set isn't directly due to the directory. On
newer ubuntu the keyring daemon uses the ipc_lock capability which
in turn prevents any FD checking via lsof. The newer ubuntu also happens
to use the new xdg directory, so we're assuming if there is a socket
there we should use it.

The lsof will eventually need to be ripped out entirely and another
solution found.